Extract getHost() and getUDN() to common handler base class.
